Refactor : 데이터 검색 로직 코드 리펙토링
데이터베이스에서 데이터를 찾았을 때, 찾지 못할 경우를 커버하기 위해, Optional을 사용하였다.

+++ src/main/java/kr/codesqaud/cafe/controller/ArticleController.java
@@ -4,7 +4,8 @@ package kr.codesqaud.cafe.controller;
 import kr.codesqaud.cafe.basic.Article;
 import kr.codesqaud.cafe.basic.Reply;
 import kr.codesqaud.cafe.basic.User;
-import kr.codesqaud.cafe.config.ConstConfig;
+import kr.codesqaud.cafe.config.ConstantConfig;
+import kr.codesqaud.cafe.exception.gobalExeption.NotFoundException;
 import kr.codesqaud.cafe.repository.ArticleRepository;
 import kr.codesqaud.cafe.repository.DataBaseRepository.DatabaseReplyRepository;
 import org.springframework.stereotype.Controller;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Controller
 @RequestMapping("/qna")
@@ -27,14 +29,14 @@ public class ArticleController {
     }
 
     @GetMapping("/create")
-    public String create(@SessionAttribute(name = ConstConfig.SESSION_ID) User user, Model model) {
+    public String create(@SessionAttribute(name = ConstantConfig.SESSION_ID) User user, Model model) {
         return "qna/create";
     }
 
     @PostMapping("/create")
     public String create(@RequestParam String title,
                          @RequestParam String contents,
-                         @SessionAttribute(name = ConstConfig.SESSION_ID) User user) {
+                         @SessionAttribute(name = ConstantConfig.SESSION_ID) User user) {
         articleRepository.save(new Article(user.getUserId(), title, contents));
         return "redirect:/qna/list";
     }
@@ -50,7 +52,11 @@ public class ArticleController {
     @GetMapping("/show/{articleId}")
     public String show(@PathVariable int articleId,
                        Model model) {
-        Article article = articleRepository.findByArticleId(articleId);
+        // TODO : 단순히 게시글을 찾는 로직이지만, 코드가 길어지는 뜻한 느낌이 있음
+        // TODO : 서비스 계층에서 처리하는 것을 고려할 예정 (예외는 컨트롤에서 하는 것이 좋다고 들었는데, 예외는 어떻게 처리?)
+        Optional<Article> articleOptional = articleRepository.findByArticleId(articleId);
+        if (articleOptional.isEmpty()) throw new NotFoundException("게시글을 찾지 못했습니다.");
+        Article article = articleOptional.get();
         List<Reply> replies = replyRepository.findByArticleId(articleId);
         model.addAttribute("article", article);
         model.addAttribute("replies", replies);
@@ -59,10 +65,16 @@ public class ArticleController {
 
     @GetMapping("/updateForm/{articleId}")
     public String update(@PathVariable int articleId,
-                         @SessionAttribute(name = ConstConfig.SESSION_ID) User user,
+                         @SessionAttribute(name = ConstantConfig.SESSION_ID) User user,
                          Model model) {
+        // TODO : 단순히 게시글을 찾는 로직이지만, 코드가 길어지는 뜻한 느낌이 있음
+        // TODO : 서비스 계층에서 처리하는 것을 고려할 예정 (예외는 컨트롤에서 하는 것이 좋다고 들었는데, 예외는 어떻게 처리?)
+        Optional<Article> articleOptional = articleRepository.findByArticleId(articleId);
+        if (articleOptional.isEmpty()) throw new NotFoundException("게시글을 찾지 못했습니다.");
+        Article article = articleOptional.get();
+
         // TODO : 검증 공부 후 다시 리펙토링 할 예정
-        Article article = articleRepository.findByArticleId(articleId);
+        // TODO : 그냥 리스트로 돌아가는 것이 아닌, 명시적인 에러처리를 해야함
         if (!article.getWriter().equals(user.getUserId())) {
             return "redirect:/qna/list";
         }
@@ -81,12 +93,19 @@ public class ArticleController {
 
     @DeleteMapping("/delete/{articleId}")
     public String delete(@PathVariable int articleId,
-                         @SessionAttribute(name = ConstConfig.SESSION_ID) User user) {
-        Article article = articleRepository.findByArticleId(articleId);
+                         @SessionAttribute(name = ConstantConfig.SESSION_ID) User user) {
+        // TODO : 단순히 게시글을 찾는 로직이지만, 코드가 길어지는 뜻한 느낌이 있음
+        // TODO : 서비스 계층에서 처리하는 것을 고려할 예정 (예외는 컨트롤에서 하는 것이 좋다고 들었는데, 예외는 어떻게 처리?)
+        Optional<Article> articleOptional = articleRepository.findByArticleId(articleId);
+        if (articleOptional.isEmpty()) throw new NotFoundException("게시글을 찾지 못했습니다");
+        Article article = articleOptional.get();
+
         // TODO : 검증 공부 후 다시 리펙토링 할 예정
+        // TODO : 그냥 리스트로 돌아가는 것이 아닌, 명시적인 에러처리를 해야함
         if (article.getWriter().equals(user.getUserId())) {
-            articleRepository.delete(articleId);
+            return "redirect:/qna/list";
         }
+        articleRepository.delete(articleId);
         return "redirect:/qna/list";
     }
